requirements.txt
docker-py==1.3.0
python-etcd==0.3.3
